requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
openai==1.57.0
langchain==0.3.7
langchain-openai==0.2.9
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
oauthlib==3.2.2
PyJWT==2.10.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
pytest==8.3.4
websockets==11.0.3